requirements.txt
streamlit==1.36.0
pandas==2.2.2
gspread==6.1.2
gspread-dataframe==3.3.1
google-auth==2.33.0
google-api-python-client==2.141.0